refactor(navbar): 내브바에서도 useAuth 쓰도록 변경 #224
