+++ src/features/front/useFront.js
@@ -1,51 +1,10 @@
 import useLinkUpStore from "../../shared/store/store";
 import { useEffect } from "react";
 import { useQuery } from "@tanstack/react-query";
-import { useNavigate, useSearchParams } from "react-router";
+import { useSearchParams } from "react-router";
 import { axiosReturnsData } from "../../shared/services/axiosInstance";
 import queryClient from "../../shared/services/queryClient";
-
-const useAuthMe = () => {
-    const navigate = useNavigate();
-    const access_token = useLinkUpStore((state) => state.access_token);
-    const setUser = useLinkUpStore((state) => state.setUser);
-
-    const endpointAuthMe = "/api/auth/me";
-    const {
-        data,
-        error: errorAuthMe,
-        isPending: isPendingAuthMe,
-    } = useQuery({
-        queryKey: [endpointAuthMe, access_token],
-        queryFn: () => axiosReturnsData("GET", endpointAuthMe),
-        refetchOnWindowFocus: false,
-        enabled: Boolean(access_token),
-    });
-
-    useEffect(() => {
-        setUser(data ?? null);
-
-        if (!data) {
-            return;
-        }
-        const user_type = data.user_type;
-        switch (user_type) {
-            case "fan":
-                break;
-            case "company":
-                navigate("/agency");
-                break;
-            case "admin":
-                navigate("/super-user");
-                break;
-            default:
-                throw new Error("---- ERROR OCCURRED: 잘못된 유저 타입:", user_type);
-        }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [data]);
-
-    return { isPendingAuthMe, errorAuthMe };
-};
+import useAuth from "../../shared/services/useAuth";
 
 const useSubscriptions = () => {
     const setArtistArray = useLinkUpStore((state) => state.setArtistArray);
@@ -103,8 +62,8 @@ const useClearCacheAfterLogout = () => {
 export const useFront = () => {
     useClearCacheAfterLogout();
     useParseTokenFromRedirectedUrl();
-    const authMeReturn = useAuthMe();
+    const authReturn = useAuth();
     const subscriptionsReturn = useSubscriptions();
 
-    return { ...authMeReturn, ...subscriptionsReturn };
+    return { ...authReturn, ...subscriptionsReturn };
 };
