+++ src/features/layout/Navbar.jsx
@@ -4,8 +4,16 @@ import CustomInput from "../../package/CustomInput";
 import { Hstack } from "../../package/layout";
 import CustomButton from "../../package/customButton/CustomButton";
 import useLinkUpStore from "../../shared/store/store";
-import useNavbar from "./useNavbar";
+import useAuth from "../../shared/services/useAuth";
 
+const DebugButton = () => {
+    const access_token = useLinkUpStore((state) => state.access_token);
+    const user = useLinkUpStore((state) => state.user);
+    const handleClick = () => {
+        console.log({ access_token, user });
+    };
+    return <CustomButton onClick={handleClick}>DEBUG</CustomButton>;
+};
 const SideSection = ({ justify, children, ...props }) => (
     <Hstack justify={justify} className={styles.sideSection} {...props}>
         {children}
@@ -14,7 +22,7 @@ const SideSection = ({ justify, children, ...props }) => (
 
 const FanMypageButton = () => {
     const navigate = useNavigate();
-    return <CustomButton onClick={() => navigate("/mypage")}>마이페이지</CustomButton>; 
+    return <CustomButton onClick={() => navigate("/mypage")}>마이페이지</CustomButton>;
 };
 const AgencyMypageButton = () => {
     const navigate = useNavigate();
@@ -44,7 +52,7 @@ const Navbar = () => {
     const navigate = useNavigate();
     const [, setSearchParams] = useSearchParams();
     const user = useLinkUpStore((state) => state.user);
-    const { logout } = useNavbar();
+    const { logout } = useAuth();
 
     const handleSearch = (keyword) => {
         const trimmed = keyword.trim();
@@ -54,6 +62,7 @@ const Navbar = () => {
     return (
         <Hstack justify="center" items="center">
             <Hstack items="center" justify="space-between" className={styles.navbar}>
+                <DebugButton />
                 <SideSection justify="start">
                     <Link to="/" className={styles.logo}>
                         Logo
@@ -65,7 +74,7 @@ const Navbar = () => {
                     onEnter={handleSearch}
                     className={styles.searchbar}
                 />
-                
+
                 <SideSection justify="end">
                     <MyButton user={user} />
                     {user && <CustomButton onClick={logout}>로그아웃</CustomButton>}
